Nombre del Proyecto: Mi Monte Fase 2 - Conciliación    * Se realiza la correccion de ugs encontradis durante revicion de servicios de pagos en linea y reporte de pagos

+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/MovimientoTransitoRepository.java
@@ -52,7 +52,7 @@ public interface MovimientoTransitoRepository extends JpaRepository<MovimientoTr
 	 */
 	@Query("SELECT mt FROM MovimientoConciliacion mc INNER JOIN MovimientoTransito mt ON mc.id = mt.id WHERE mc.idConciliacion = :folio AND mt.id = :idMovimiento ")
 	public MovimientoTransito findByIdFolioAndIdMovimiento(@Param("folio") final Integer folio,
-			@Param("idMovimiento") final long idMovimiento);
+			@Param("idMovimiento") final Integer idMovimiento);
 
 	/**
 	 * Regresa una lista de MovimientoTransito por folio e ids
